requirements.txt
picamera2
pyzmq